+++ lib/ansible/executor/task_result.py
@@ -8,7 +8,7 @@ __metaclass__ = type
 from copy import deepcopy
 
 from ansible.parsing.dataloader import DataLoader
-from ansible.vars.manager import strip_internal_keys
+from ansible.vars.clean import strip_internal_keys
 
 _IGNORE = ('failed', 'skipped')
