+++ lib/ansible/plugins/action/package.py
@@ -43,9 +43,9 @@ class ActionModule(ActionBase):
         if module == 'auto':
             try:
                 if self._task.delegate_to:  # if we delegate, we should use delegated host's facts
-                    module = self._templar.template("{{hostvars['%s']['ansible_pkg_mgr']}}" % self._task.delegate_to)
+                    module = self._templar.template("{{hostvars['%s']['ansible_facts']['pkg_mgr']}}" % self._task.delegate_to)
                 else:
-                    module = self._templar.template('{{ansible_pkg_mgr}}')
+                    module = self._templar.template('{{ansible_facts.pkg_mgr}}')
             except:
                 pass  # could not get it from template!
