+++ lib/ansible/plugins/action/service.py
@@ -42,9 +42,9 @@ class ActionModule(ActionBase):
         if module == 'auto':
             try:
                 if self._task.delegate_to:  # if we delegate, we should use delegated host's facts
-                    module = self._templar.template("{{hostvars['%s']['ansible_service_mgr']}}" % self._task.delegate_to)
+                    module = self._templar.template("{{hostvars['%s']['ansible_facts']['service_mgr']}}" % self._task.delegate_to)
                 else:
-                    module = self._templar.template('{{ansible_service_mgr}}')
+                    module = self._templar.template('{{ansible_facts.service_mgr}}')
             except:
                 pass  # could not get it from template!
