+++ lib/ansible/plugins/callback/__init__.py
@@ -29,7 +29,7 @@ from ansible import constants as C
 from ansible.plugins import AnsiblePlugin
 from ansible.module_utils._text import to_text
 from ansible.utils.color import stringc
-from ansible.vars.manager import strip_internal_keys
+from ansible.vars.clean import strip_internal_keys
 
 try:
     from __main__ import display as global_display
